Move around a bunch of webpack files
