+++ lib/logger.js
@@ -2,7 +2,7 @@
 
 const chalk = require('chalk');
 const timestamp = require('time-stamp');
-const { WebpackCompileError } = require('./webpack-helpers');
+const { WebpackCompileError } = require('./webpack-compiler');
 
 function stampedMessage(message) {
   const stamp = `[${chalk.grey(timestamp('HH:mm:ss'))} ${chalk.cyan(
