+++ lib/webpack-compiler/create-webpack-compiler.js
@@ -2,8 +2,8 @@
 
 const webpack = require('webpack');
 
+const createWebpackConfig = require('../webpack-config');
 const normalizeWebpackError = require('./normalize-webpack-errors');
-const createWebpackConfig = require('./create-webpack-config');
 
 module.exports = function createWebpackCompiler(urc) {
   const webpackConfig = createWebpackConfig(urc);
