+++ lib/webpack-compiler/index.js
