+++ lib/webpack-compiler/normalize-webpack-errors.js
