+++ lib/webpack-compiler/webpack-compile-error.js
