+++ lib/webpack-compiler/webpack-promise.js
@@ -1,8 +1,9 @@
 'use strict';
 
 const webpack = require('webpack');
+
+const createWebpackConfig = require('../webpack-config/create-webpack-config');
 const normalizeWebpackError = require('./normalize-webpack-errors');
-const createWebpackConfig = require('./create-webpack-config');
 
 function webpackPromise(urc) {
   const webpackConfig = createWebpackConfig(urc);
