+++ lib/webpack-compiler/write-webpack-stats.js
