+++ lib/webpack-config/create-webpack-config.js
@@ -5,8 +5,8 @@ const resolvePkg = require('resolve-pkg');
 
 const compact = require('../utils/compact-array');
 const safeWebpackUglifier = require('./safe-webpack-uglifier');
-const webpackPlugins = require('./webpack-plugins');
-const webpackLoaders = require('./webpack-loaders');
+const webpackPluginsList = require('./plugins-list');
+const webpackLoadersList = require('./loaders-list');
 
 const webpackEntry = urc => {
   const jsEntryDir = path.dirname(urc.jsEntry);
@@ -49,7 +49,7 @@ module.exports = function createWebpackConfig(urc) {
     module: {
       rules: [
         {
-          oneOf: webpackLoaders(urc)
+          oneOf: webpackLoadersList(urc)
         }
       ]
     },
@@ -90,7 +90,7 @@ module.exports = function createWebpackConfig(urc) {
         : `${publicAssetsPath}/js/[name].chunk.js`
     },
     performance: false,
-    plugins: webpackPlugins(urc),
+    plugins: webpackPluginsList(urc),
     resolve: {
       alias: {
         // This is added to make Underreact symlinking work with npm <=3,
