+++ lib/webpack-config/generate-html-template.js
@@ -2,9 +2,11 @@
 const path = require('path');
 const sourceMappingURL = require('source-map-url');
 
-module.exports = { generateTemplate };
-
-function generateTemplate({ urc, webpackCompilation, publicPath }) {
+module.exports = function generateHtmlTemplate({
+  urc,
+  webpackCompilation,
+  publicPath
+}) {
   return Promise.resolve(
     // TOFIX: We need to make sure the urc.htmlSource returns a valid html
     urc.htmlSource({
@@ -30,7 +32,7 @@ function generateTemplate({ urc, webpackCompilation, publicPath }) {
         })}`
       )
   );
-}
+};
 
 // Returns a script which conditionally loads the `polyfill.js` file
 // if and only if the browser fails the `urc.modernBrowserTest`.
