+++ lib/webpack-config/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('./create-webpack-config');
