+++ lib/webpack-config/loaders-list.js
@@ -7,7 +7,7 @@ const autoprefixer = require('autoprefixer');
 
 const compact = require('../utils/compact-array');
 
-module.exports = function webpackLoaders(urc) {
+module.exports = function webpackLoadersList(urc) {
   const babelConfig = createBabelLoader(urc);
   const publicAssetsPath = urc.publicAssetsPath.replace(/^\//, '');
