+++ lib/webpack-config/plugins-list.js
@@ -8,11 +8,11 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 const { WEBPACK_ASSETS_BASENAME } = require('../constants');
 const compact = require('../utils/compact-array');
-const ExcludeChunksPlugin = require('./exclude-chunks-plugin');
-const ModuleNotFoundPlugin = require('./module-not-found-plugin');
-const { generateTemplate } = require('./helpers');
+const ExcludeChunksPlugin = require('../webpack-plugins/exclude-chunks-plugin');
+const ModuleNotFoundPlugin = require('../webpack-plugins/module-not-found-plugin');
+const generateHtmlTemplate = require('./generate-html-template');
 
-module.exports = function getWebpackPlugins(urc) {
+module.exports = function webpackPluginsList(urc) {
   const publicAssetsPath = urc.publicAssetsPath.replace(/^\//, '');
   return compact([
     new ModuleNotFoundPlugin(urc.rootDirectory),
@@ -45,7 +45,7 @@ module.exports = function getWebpackPlugins(urc) {
             minifyURLs: true
           },
       templateContent: ({ compilation, htmlWebpackPlugin }) =>
-        generateTemplate({
+        generateHtmlTemplate({
           urc,
           webpackCompilation: compilation,
           publicPath: htmlWebpackPlugin.files.publicPath
